RAG_QA/requirements.txt
streamlit==1.37.0
python-dotenv==1.0.1

# ---- LangChain ecosystem ----
langchain==0.1.16
langchain-core==0.1.52
langchain-community==0.0.33
langchain-text-splitters==0.0.2
langchain-groq==0.1.3
langchain-google-genai==0.0.8

# ---- Vector DB & PDF ----
faiss-cpu==1.7.4
pypdf==4.2.0

# ---- PyTorch for Windows CPU ----
torch==2.1.2









